requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
WTForms==3.0.1
Flask_WTF==1.2.1
Werkzeug==3.0.0
python-dotenv==1.0.1
wtforms[email]
bootstrap-flask==2.2.0